[CORE] Switched to using 'Int' as the default data type for shapes, rather than 'Long', which can provide a significant performance boost when using GPUs.

+++ modules/api/src/main/scala/org/platanios/tensorflow/api/implicits/Implicits.scala
@@ -82,11 +82,11 @@ private[api] trait Implicits
     outputFromSupportedType[Double](value)
   }
 
-  implicit def shapeToTensor(shape: Shape): Tensor[Long] = {
+  implicit def shapeToTensor(shape: Shape): Tensor[Int] = {
     shape.toTensor
   }
 
-  implicit def shapeToOutput(shape: Shape): Output[Long] = {
+  implicit def shapeToOutput(shape: Shape): Output[Int] = {
     shape.toOutput
   }
